backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32
PyMySQL==1.1.1
pyodbc==5.1.0
bcrypt==4.1.3
pandas==2.2.2
